image_filter_app/requirements.txt
Flask
Pillow
matplotlib
numpy
